+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Replace `android-database-sqlcipher` library with `sqlcipher-android`
 - Replace Flipper with Chucker for network inspection
 - Update statin risk threshold to 20% for sri Lanka
+- Update smoking dialog to a multi-choice tobacco dialog
 
 ## 2025.05.20